[env] normalize

+++ services/worker/entry.sh
@@ -1,9 +1,14 @@
 #!/bin/sh
 
-while true; do
-  sleep 1800
-  killall chrome
-done &
-
 echo "starting worker..."
-xvfb-run node --expose-gc ./dist/index.js
+
+if [ "$DISH_ENV" = "development" ]; then
+  export CLEAR_JOBS
+  ../../dishctl.sh run 'node dist/index.js'
+else
+  while true; do
+    sleep 1800
+    killall chrome
+  done &
+  xvfb-run node --expose-gc ./dist/index.js
+fi
